@@ -6,7 +6,8 @@
   "scripts": {
     "compile": "gulp compile",
     "test": "gulp test",
-    "prepublish": "gulp test",
+    "lint": "gulp lint",
+    "prepublishOnly": "gulp test",
     "functional": "gulp functional-test",
     "browserify": "gulp browserify",
     "prepare": "npm run compile"
@@ -30,36 +31,36 @@
   },
   "homepage": "https://github.com/minio/minio-js#readme",
   "dependencies": {
-    "async": "^1.5.2",
-    "block-stream2": "^1.0.0",
-    "concat-stream": "^1.4.8",
-    "es6-error": "^2.0.2",
+    "async": "^3.1.0",
+    "block-stream2": "^2.0.0",
+    "es6-error": "^4.1.1",
     "json-stream": "^1.0.0",
     "lodash": "^4.14.2",
     "mime-types": "^2.1.14",
     "mkdirp": "^0.5.1",
     "querystring": "0.2.0",
-    "through2": "^0.6.5",
-    "uuid": "^3.1.0",
+    "through2": "^3.0.1",
     "xml": "^1.0.0",
     "xml2js": "^0.4.15"
   },
   "devDependencies": {
-    "browserify": "^12.0.1",
-    "chai": "^3.5.0",
-    "eslint": "^4.1.1",
-    "gulp": "^3.9.0",
-    "gulp-babel": "^5.2.1",
-    "gulp-eslint": "^4.0.0",
-    "gulp-mocha": "^2.1.0",
-    "gulp-notify": "^2.2.0",
-    "gulp-sourcemaps": "^1.5.2",
-    "mocha": "^2.3.2",
+    "@babel/core": "^7.5.5",
+    "@babel/preset-env": "^7.5.5",
+    "babelify": "^10.0.0",
+    "browserify": "^16.3.0",
+    "chai": "^4.2.0",
+    "eslint": "^6.1.0",
+    "gulp": "^4.0.2",
+    "gulp-babel": "^8.0.0",
+    "gulp-eslint": "^6.0.0",
+    "gulp-mocha": "^6.0.0",
+    "gulp-sourcemaps": "^2.6.5",
+    "mocha": "^6.2.0",
     "mocha-steps": "^1.1.0",
-    "nock": "^2.12.0",
-    "rewire": "^2.3.3",
-    "source-map-support": "^0.4.12",
-    "superagent": "^3.7.0"
+    "nock": "^10.0.6",
+    "source-map-support": "^0.5.13",
+    "superagent": "^5.1.0",
+    "uuid": "^3.3.2"
   },
   "keywords": [
     "api",
